Extract tests for isURIError to own file
