+++ lib/referee.test.js
@@ -4,43 +4,6 @@ var referee = require("../lib/referee");
 var testHelper = require("./test-helper");
 var sinon = require("sinon");
 
-testHelper.assertionTests("assert", "isURIError", function(
-    pass,
-    fail,
-    msg,
-    error
-) {
-    function captureArgs() {
-        return arguments;
-    }
-
-    fail("for Error", new Error("not pie"));
-    pass("for URIError", new URIError("apple pie"));
-    fail("for string", "apple pie");
-    fail("for array", []);
-    fail("for object", {});
-    fail("for arguments", captureArgs());
-    msg(
-        "fail with descriptive message",
-        "[assert.isURIError] Expected {  } to be a URIError",
-        {}
-    );
-    msg(
-        "fail with custom message",
-        "[assert.isURIError] Nope: Expected {  } to be a URIError",
-        {},
-        "Nope"
-    );
-    error(
-        "for object",
-        {
-            code: "ERR_ASSERTION",
-            operator: "assert.isURIError"
-        },
-        {}
-    );
-});
-
 testHelper.assertionTests("assert", "isUint16Array", function(
     pass,
     fail,
